# qooauth/api/openapi/auth-api.yaml
openapi: "3.0.3"
info:
  title: QooAuth API
  description: QooBot 统一身份认证基础设施 API
  version: "0.1.0"
  contact:
    name: QooBot Team
    url: https://github.com/qoobot/qooauth

servers:
  - url: https://auth.qoobot.com/api/v1
    description: 生产环境
  - url: http://localhost:8080/api/v1
    description: 本地开发

paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: 用户注册
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '409':
          $ref: '#/components/responses/Conflict'

  /auth/login:
    post:
      tags: [Authentication]
      summary: 密码登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/refresh:
    post:
      tags: [Authentication]
      summary: 刷新 Token
      operationId: refreshToken
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token 刷新成功

  /auth/logout:
    post:
      tags: [Authentication]
      summary: 登出
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功

  /users/me:
    get:
      tags: [Users]
      summary: 获取当前用户资料
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 用户资料

  /devices:
    get:
      tags: [Devices]
      summary: 设备列表
      operationId: listDevices
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 设备列表

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        display_name:
          type: string
        locale:
          type: string
        timezone:
          type: string

    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/UserProfile'
        requires_2fa:
          type: boolean

    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string

    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object

  responses:
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'